Use uint32_t instead of size_t
git-svn-id: http://svn.macosforge.org/repository/mdnsresponder/trunk@4505 214c2c4a-bf3b-4dcf-9390-e4dd3010487d

+++ mDNSShared/dns_sd.h
@@ -510,9 +510,9 @@ typedef int32_t  DNSServiceErrorType;
 
 DNSServiceErrorType DNSSD_API DNSServiceGetProperty
     (
-    char    *attr,      /* Requested attribute (DNSSD_GetProperty_DaemonVersion) */
-    void    *result,    /* Pointer to place to store result */
-    size_t  size        /* size of result location */
+    char     *attr,      /* Requested attribute (DNSSD_GetProperty_DaemonVersion) */
+    void     *result,    /* Pointer to place to store result */
+    uint32_t size        /* size of result location */
     );
